+++ multisrc/overrides/mangareader/comickiba/src/ManhuagoldFilters.kt
@@ -1,0 +1,94 @@
+package eu.kanade.tachiyomi.extension.en.comickiba
+
+import eu.kanade.tachiyomi.multisrc.mangareader.MangaReader.UriMultiSelectFilter
+import eu.kanade.tachiyomi.multisrc.mangareader.MangaReader.UriPartFilter
+
+class StatusFilter : UriPartFilter(
+    "Status",
+    "status",
+    arrayOf(
+        Pair("All", ""),
+        Pair("Completed", "completed"),
+        Pair("OnGoing", "on-going"),
+        Pair("On-Hold", "on-hold"),
+        Pair("Canceled", "canceled"),
+    ),
+)
+
+class GenreFilter : UriMultiSelectFilter(
+    "Genres",
+    "genres",
+    arrayOf(
+        Pair("Action", "37"),
+        Pair("Adaptation", "19"),
+        Pair("Adult", "5310"),
+        Pair("Adventure", "38"),
+        Pair("Aliens", "5436"),
+        Pair("Animals", "1552"),
+        Pair("Award Winning", "39"),
+        Pair("Comedy", "202"),
+        Pair("Comic", "287"),
+        Pair("Cooking", "277"),
+        Pair("Crime", "2723"),
+        Pair("Delinquents", "4438"),
+        Pair("Demons", "379"),
+        Pair("Drama", "3"),
+        Pair("Ecchi", "17"),
+        Pair("Fantasy", "197"),
+        Pair("Full Color", "13"),
+        Pair("Gender Bender", "221"),
+        Pair("Genderswap", "2290"),
+        Pair("Ghosts", "2866"),
+        Pair("Gore", "42"),
+        Pair("Harem", "222"),
+        Pair("Historical", "4"),
+        Pair("Horror", "5"),
+        Pair("Isekai", "259"),
+        Pair("Josei", "292"),
+        Pair("Loli", "5449"),
+        Pair("Long Strip", "7"),
+        Pair("Magic", "272"),
+        Pair("Manhwa", "266"),
+        Pair("Martial Arts", "40"),
+        Pair("Mature", "5311"),
+        Pair("Mecha", "2830"),
+        Pair("Medical", "1598"),
+        Pair("Military", "43"),
+        Pair("Monster Girls", "2307"),
+        Pair("Monsters", "298"),
+        Pair("Music", "3182"),
+        Pair("Mystery", "6"),
+        Pair("Office Workers", "14"),
+        Pair("Official Colored", "1046"),
+        Pair("Philosophical", "2776"),
+        Pair("Post-Apocalyptic", "1059"),
+        Pair("Psychological", "493"),
+        Pair("Reincarnation", "204"),
+        Pair("Reverse", "280"),
+        Pair("Reverse Harem", "199"),
+        Pair("Romance", "186"),
+        Pair("School Life", "601"),
+        Pair("Sci-Fi", "1845"),
+        Pair("Sexual Violence", "731"),
+        Pair("Shoujo", "254"),
+        Pair("Slice of Life", "10"),
+        Pair("Sports", "4066"),
+        Pair("Superhero", "481"),
+        Pair("Supernatural", "198"),
+        Pair("Survival", "44"),
+        Pair("Thriller", "1058"),
+        Pair("Time Travel", "299"),
+        Pair("Tragedy", "41"),
+        Pair("Video Games", "1846"),
+        Pair("Villainess", "278"),
+        Pair("Virtual Reality", "1847"),
+        Pair("Web Comic", "12"),
+        Pair("Webtoon", "279"),
+        Pair("Webtoons", "267"),
+        Pair("Wuxia", "203"),
+        Pair("Yaoi", "18"),
+        Pair("Yuri", "11"),
+        Pair("Zombies", "1060"),
+    ),
+    ",",
+)
